UI and Master Changes
UI now displays the status of buildings.
UI now allows the user to continue seeing the previously selected action
information after moving away from it.
More.

+++ MapObjects/Buildings/Base.cs
@@ -41,7 +41,7 @@ public class Base : Building {
 		//Do nothing, the base cannot be killed
 	}
 	
-	public void DepositMass(int amount) {
+	public void DepositMass(float amount) {
 		manager.AddMass(amount);
 	}
